+++ test/fixtures/test.manifest-fixture.json
